working implementation for callback functions
